[refactor] boards 리팩토링
