+++ src/pages/boards/[board_type]/posts/index.tsx
@@ -1,0 +1,18 @@
+/* eslint-disable react/jsx-props-no-spreading */
+
+import CommonCenterWrapper from '@src/components/common/Layout/CommonCenterWrapper';
+import PostForm from '@src/components/pages/posts/PostForm';
+
+/**
+ *  글 작성 페이지
+ */
+
+function PostPage() {
+  return (
+    <CommonCenterWrapper>
+      <PostForm />
+    </CommonCenterWrapper>
+  );
+}
+
+export default PostPage;
